Created Audio Controller, Replaced Existing Audio Files with Bitcrushed Ones, and Implemented Player and Enemy SFX into the game
New Audio Controller node can play sounds in any script/file and is automatically loaded upon startup. Eventually we should probably move the main music loop here, time permitting. I implemented the SFX for the following:
- Laser Gunshot
- Player Getting Hurt
- Player Dying
- Enemy Getting Hurt
- Enemy Dying

Note: I also added a bitcrushed version of the main theme, in case we want to replace it with the old one. I haven't yet but we can do that if need be.