Created Game Over Screen and Modified Enemy AI
Created and implemented the Game Over screen, including the art. I also modified the Enemy AI to stop attacking once the player has died. The issue was because upon dying, I changed the scene to "Game Over". However, the enemy AI was still attacking and in the main scene. When it tried accessing the tree, it crashed the game before we could even get to the game over screen. Fix was to add a hit counter, though an alternative fix could be to move the timer to not use the tree, if possible.